# Conflicts:
#	src/main/java/com/cegeka/academy/repository/GroupUserRoleRepository.java

+++ src/main/java/com/cegeka/academy/web/rest/InvitationController.java
@@ -1,13 +1,18 @@
 package com.cegeka.academy.web.rest;
 
 import com.cegeka.academy.domain.Invitation;
+import com.cegeka.academy.repository.InvitationRepository;
 import com.cegeka.academy.service.dto.InvitationDTO;
 import com.cegeka.academy.service.invitation.InvitationService;
-import com.cegeka.academy.service.invitation.ValidationAccessService;
+import com.cegeka.academy.service.serviceValidation.ValidationAccessService;
+import com.cegeka.academy.web.rest.errors.NotFoundException;
+import com.cegeka.academy.web.rest.errors.UnauthorizedUserException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("invitation")
@@ -15,12 +20,15 @@ public class InvitationController {
 
     private final InvitationService invitationService;
 
+    private final InvitationRepository invitationRepository;
+
     private final ValidationAccessService validationAccessService;
 
     @Autowired
-    public InvitationController(InvitationService invitationService, ValidationAccessService validationAccessService) {
+    public InvitationController(InvitationService invitationService, ValidationAccessService validationAccessService, InvitationRepository invitationRepository) {
         this.invitationService = invitationService;
         this.validationAccessService = validationAccessService;
+        this.invitationRepository = invitationRepository;
     }
 
     @GetMapping("/all")
@@ -30,28 +38,82 @@ public class InvitationController {
     }
 
     @PostMapping
-    public void saveInvitation(@RequestBody Invitation newInvitation){
+    public void saveInvitation(@Valid @RequestBody Invitation newInvitation) {
 
          invitationService.saveInvitation(newInvitation);
     }
 
     @PutMapping
-    public void replaceInvitation(@RequestBody Invitation newInvitation){
+    public void replaceInvitation(@RequestBody Invitation newInvitation) throws NotFoundException {
 
-        if(validationAccessService.verifyUserAccessForInvitationEntity(newInvitation.getId()))
-        {
-            invitationService.updateInvitation(newInvitation);
+        Optional<Invitation> updateInvitation = invitationRepository.findById(newInvitation.getId());
 
+        if (updateInvitation.isPresent()) {
+
+            if (validationAccessService.verifyUserAccessForInvitationEntity(newInvitation.getId())) {
+
+                invitationService.updateInvitation(newInvitation);
+
+            } else {
+
+                throw new UnauthorizedUserException();
+            }
+
+        } else {
+
+            throw new NotFoundException();
         }
     }
 
     @DeleteMapping("/{id}")
-    void deleteInvitation(@PathVariable Long id) {
+    public void deleteInvitation(@PathVariable Long id) throws NotFoundException {
 
-        if(validationAccessService.verifyUserAccessForInvitationEntity(id))
-        {
-            invitationService.deleteInvitationById(id);
+        Optional<Invitation> deleteInvitation = invitationRepository.findById(id);
+
+        if (deleteInvitation.isPresent()) {
+
+            if (validationAccessService.verifyUserAccessForInvitationEntity(id)) {
+
+                invitationService.deleteInvitationById(id);
+
+            } else {
+
+                throw new UnauthorizedUserException();
+            }
+        } else {
+
+            throw new NotFoundException();
         }
     }
 
+    @GetMapping("/pending/{id}")
+    public List<InvitationDTO> getPendingInvitationsByUserId(@PathVariable Long userId) {
+
+        return invitationService.getPendingInvitationsByUserId(userId);
+    }
+
+    @PutMapping("/accept")
+    public void acceptInvitation(@RequestBody Invitation newInvitation) {
+
+        if (validationAccessService.verifyUserAccessForInvitationEntity(newInvitation.getId())) {
+            invitationService.acceptInvitation(newInvitation);
+
+        }
+    }
+
+    @PutMapping("/reject")
+    public void rejectInvitation(@RequestBody Invitation newInvitation) {
+
+        if (validationAccessService.verifyUserAccessForInvitationEntity(newInvitation.getId())) {
+            invitationService.rejectInvitation(newInvitation);
+
+        }
+    }
+
+    @PostMapping("/send/{idGroup}")
+    public void sendGroupInvitationsToPrivateEvents(@PathVariable Long idGroup,
+                                                    @Valid @RequestBody Invitation newInvitation) throws NotFoundException {
+
+        invitationService.sendGroupInvitationsToPrivateEvents(idGroup, newInvitation);
+    }
 }
